pyproject.toml
[tool.poetry]
name = "olx"
version = "0.1.0"
description = ""
authors = ["dinmukhammed <arystambekdimash005@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.2"
alembic = "^1.12.0"
uvicorn = "^0.23.2"
pydantic = "^2.4.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
sqlalchemy = "^2.0.21"
pydantic-settings = "^2.0.3"
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
